Add ActiveModel::Attributes#==.
I want to be able to compare identity checks with == method.
Because I want to compare attribute identities,
but currently I have to go through the attribute method.

I wanted to use AttributeSet#== for the comparison to avoid creating
unnecessary objects, but since it does not return true if the value
before type cast are different, I used Attributes#attributes instead.